+++ phoenicis-cli/src/main/java/org/phoenicis/cli/setupwindow/SetupUiCliImplementation.java
@@ -32,12 +32,12 @@ import java.util.List;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
 
-class SetupUiCLIImplementation implements SetupUi {
+class SetupUiCliImplementation implements SetupUi {
     private final String title;
     private final boolean interactive;
     private final boolean verbose;
 
-    SetupUiCLIImplementation(String title, boolean interactive, boolean verbose) {
+    SetupUiCliImplementation(String title, boolean interactive, boolean verbose) {
         this.title = title;
         this.interactive = interactive;
         this.verbose = verbose;
